Adds iOS 6 functionality, placeholder sections and tweaks minor UI errors.
The app nows determines if the user has iOS 6 or iOS 7 installed. The app also
allows for all buttons on the home screen to be clicked, even though most
of the sections contain no data. The only section containing any text
(besides Contacts) is the “Agenda” section, containing a table view with
placeholder data relating to conference activities.